Move app service common attributes from webapp runtime builder to base runtime builder
