+++ azure-webapp-maven-plugin/src/main/java/com/microsoft/azure/maven/webapp/handlers/runtime/WebAppRuntimeHandler.java
@@ -39,85 +39,13 @@ public abstract class WebAppRuntimeHandler extends BaseRuntimeHandler<WebApp> {
     protected String registryUrl;
     protected Log log;
 
-    public abstract static class Builder<T extends Builder<T>> {
-        private RuntimeStack runtime;
+    public abstract static class Builder<T extends WebAppRuntimeHandler.Builder<T>> extends BaseRuntimeHandler.Builder<T> {
+        protected RuntimeStack runtime;
         protected JavaVersion javaVersion;
         protected WebContainer webContainer;
-        private String appName;
-        private String resourceGroup;
-        private Region region;
-        private PricingTier pricingTier;
-        private String servicePlanName;
-        private String servicePlanResourceGroup;
-        private Azure azure;
-        private Settings settings;
-        protected String image;
-        protected String serverId;
-        protected String registryUrl;
-        private Log log;
 
         public T runtime(final RuntimeStack value) {
             this.runtime = value;
-            return self();
-        }
-
-        public T appName(final String value) {
-            this.appName = value;
-            return self();
-        }
-
-        public T resourceGroup(final String value) {
-            this.resourceGroup = value;
-            return self();
-        }
-
-        public T region(final Region value) {
-            this.region = value;
-            return self();
-        }
-
-        public T pricingTier(final PricingTier value) {
-            this.pricingTier = value;
-            return self();
-        }
-
-        public T servicePlanName(final String value) {
-            this.servicePlanName = value;
-            return self();
-        }
-
-        public T servicePlanResourceGroup(final String value) {
-            this.servicePlanResourceGroup = value;
-            return self();
-        }
-
-        public T azure(final Azure value) {
-            this.azure = value;
-            return self();
-        }
-
-        public T mavenSettings(final Settings value) {
-            this.settings = value;
-            return self();
-        }
-
-        public T log(final Log value) {
-            this.log = value;
-            return self();
-        }
-
-        public T image(final String value) {
-            this.image = value;
-            return self();
-        }
-
-        public T serverId(final String value) {
-            this.serverId = value;
-            return self();
-        }
-
-        public T registryUrl(final String value) {
-            this.registryUrl = value;
             return self();
         }
 
@@ -155,21 +83,10 @@ public abstract class WebAppRuntimeHandler extends BaseRuntimeHandler<WebApp> {
     protected abstract OperatingSystem getAppServicePlatform();
 
     protected WebAppRuntimeHandler(Builder<?> builder) {
+        super(builder);
         this.runtime = builder.runtime;
         this.javaVersion = builder.javaVersion;
         this.webContainer = builder.webContainer;
-        this.appName = builder.appName;
-        this.resourceGroup = builder.resourceGroup;
-        this.region = builder.region;
-        this.pricingTier = builder.pricingTier;
-        this.servicePlanName = builder.servicePlanName;
-        this.servicePlanResourceGroup = builder.servicePlanResourceGroup;
-        this.azure = builder.azure;
-        this.settings = builder.settings;
-        this.image = builder.image;
-        this.serverId = builder.serverId;
-        this.registryUrl = builder.registryUrl;
-        this.log = builder.log;
     }
 
     protected AppServicePlan createOrGetAppServicePlan() throws MojoExecutionException {
